pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django backend for the Tortoise Community website."
authors = ["ryuga <vaishnav2353@gmail.com>"]
license = "MIT LICENSE"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-hosts = "^4.0"
Pillow = "^8.4.0"
httpx = "^0.20.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
python-decouple = "^3.5"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
